[ETCM-126] add helper functions newJsonRpcRequest newJsonRpcController and simplify tests using them
